requirements.txt
selenium==4.3.0
pandas==1.4.3
bs4==0.0.1
python-dotenv==0.20.0